+++ src/main/java/com/github/argon4w/acceleratedrendering/core/programs/overrides/IUploadingOverride.java
@@ -3,6 +3,6 @@ package com.github.argon4w.acceleratedrendering.core.programs.overrides;
 public non-sealed interface IUploadingOverride extends IProgramOverride {
 
 	void	uploadMeshInfo		(long	meshInfoAddress,	int	meshInfoIndex);
-	int		dispatchUploading	(int	meshCount,			int	meshSize, long meshOffset);
+	int		dispatchUploading	(int	meshCount,			int	meshSize, int meshOffset);
 	long	getMeshInfoSize		();
 }
